Move serving test to core
